refactor: verifyImageFile util method to FileUtils class
