+++ archive-application/src/main/java/site/archive/service/archive/ArchiveImageService.java
@@ -2,25 +2,12 @@ package site.archive.service.archive;
 
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.Arrays;
-
-import static org.springframework.http.MediaType.IMAGE_GIF_VALUE;
-import static org.springframework.http.MediaType.IMAGE_JPEG_VALUE;
-import static org.springframework.http.MediaType.IMAGE_PNG_VALUE;
-
 public interface ArchiveImageService {
 
     String ARCHIVE_IMAGE_DIRECTORY = "images/";
     String USER_PROFILE_IMAGE_DIRECTORY = "user-profile/";
     String BANNER_MAIN_IMAGE_DIRECTORY = "banner/main/";
     String BANNER_SUMMARY_IMAGE_DIRECTORY = "banner/summary/";
-
-    default void verifyImageFile(final MultipartFile imageFile) {
-        if (!Arrays.asList(IMAGE_PNG_VALUE, IMAGE_GIF_VALUE, IMAGE_JPEG_VALUE)
-                   .contains(imageFile.getContentType())) {
-            throw new IllegalStateException("FIle uploaded is not an image");
-        }
-    }
 
     String upload(String directory, MultipartFile imageFile);
