+++ archive-web/src/main/java/site/archive/web/api/v2/BannerControllerV2.java
@@ -16,6 +16,7 @@ import site.archive.domain.banner.BannerType;
 import site.archive.dto.v2.BannerListResponseDto;
 import site.archive.service.archive.ArchiveImageService;
 import site.archive.service.banner.BannerService;
+import site.archive.web.config.security.util.FileUtils;
 
 import static site.archive.service.archive.ArchiveImageService.BANNER_MAIN_IMAGE_DIRECTORY;
 import static site.archive.service.archive.ArchiveImageService.BANNER_SUMMARY_IMAGE_DIRECTORY;
@@ -65,7 +66,7 @@ public class BannerControllerV2 {
     }
 
     private String imageUploadAndGetUri(String directory, MultipartFile imageFile) {
-        imageService.verifyImageFile(imageFile);
+        FileUtils.verifyImageFile(imageFile);
         return imageService.upload(directory, imageFile);
     }
